# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple rate limiting and synchronization service for distributed systems",
        "title": "Bouncer API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.6"
    },
    "host": "localhost:5505",
    "basePath": "/",
    "paths": {
...
        "/semaphore/{name}": {
            "delete": {
                "description": "Remove a semaphore and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Semaphore"
                ],
                "summary": "Delete a semaphore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Semaphore name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Semaphore deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - semaphore not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/semaphore/{name}/acquire": {
            "get": {
                "description": "Acquire a semaphore lock.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Semaphore"
                ],
                "summary": "Acquire a semaphore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Semaphore name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Semaphore size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": -1,
                        "description": "Maximum wait time",
                        "name": "maxwait",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 60000,
                        "description": "Expiration time",
                        "name": "expires",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The semaphore release key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - semaphore not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "408": {
                        "description": "Request Timeout - `maxWait` exceeded",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/semaphore/{name}/release": {
            "get": {
                "description": "Release a semaphore lock",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Semaphore"
                ],
                "summary": "Release a semaphore",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Semaphore name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Semaphore size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Release key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Semaphore released successfully"
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - semaphore not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Conflict - key is invalid or already released",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tokenbucket/{name}": {
            "delete": {
                "description": "Remove a token bucket and clean up its resources",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "TokenBucket"
                ],
                "summary": "Delete a token bucket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token bucket name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Token bucket deleted successfully"
                    },
                    "404": {
                        "description": "Not Found - token bucket not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tokenbucket/{name}/acquire": {
            "get": {
                "description": "Every `interval` milliseconds, the bucket is refilled with `size` tokens.\nEach acquire request takes one token out of the bucket, or waits up to `maxWait` milliseconds for a token to be available.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "TokenBucket"
                ],
                "summary": "Acquire a token from a token bucket",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token bucket name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Bucket size",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1000,
                        "description": "Refill interval",
                        "name": "interval",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": -1,
                        "description": "Maximum wait time",
                        "name": "maxwait",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Token acquired successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request - invalid parameters",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found - token bucket not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "408": {
                        "description": "Request Timeout - `maxWait` exceeded",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}